Add files via upload
In order to minimize energy consumption of sensor nodes in wireless sensor networks, this paper presents an Energy Efficient Sleep-Scheduled Tree-Based Routing Protocol (EESSTBRP) that modifies the chain formation in PEGASIS to create a set of paired and unpaired nodes in the network based on a distance and sensing range threshold. The paired nodes switch between active and sleep modes so as to remove redundant data and save battery power. To minimize energy consumption as nodes switching between the modes, this scheme considers the transitioning to be done based on a point of near depletion of the nodes’ residual energy. To further reduce energy consumption during transmission, this protocol utilizes prim’s minimum spanning tree mechanism to route data from the active nodes to the Base Station (BS). Simulation results show that this proposed mechanism can significantly improve network lifetime in comparison to PEGASIS.   

The attached file, Routing_algorithm.m is matlab code for the proposed routing algorithm. Topology.m is the network layout. 17-0104E_publish2019110.pdf is the paper published in the Journal of  System Simulation (www.china-simulation.com)